+++ batch_command_get_reflect.go
@@ -36,6 +36,7 @@ func parseBatchObject(
 	generation uint32,
 	expiration uint32,
 ) error {
+	supportsFloat := cmd.node.cluster.supportsFloat.Get()
 	if opCount > 0 {
 		rv := *cmd.objects[offset]
 
@@ -81,7 +82,7 @@ func parseBatchObject(
 			if err != nil {
 				return err
 			}
-			if err := setObjectField(mappings, iobj, name, value); err != nil {
+			if err := setObjectField(mappings, iobj, name, value, supportsFloat); err != nil {
 				return err
 			}
 		}