+++ client.go
@@ -380,8 +380,6 @@ func (clnt *Client) BatchGetHeader(policy *BatchPolicy, keys []*Key) ([]*Record,
 // An example would be to add an integer value to an existing record and then
 // read the result, all in one database call.
 //
-// Write operations are always performed first, regardless of operation order
-// relative to read operations.
 // If the policy is nil, the default relevant policy will be used.
 func (clnt *Client) Operate(policy *WritePolicy, key *Key, operations ...*Operation) (*Record, error) {
 	policy = clnt.getUsableWritePolicy(policy)
@@ -938,8 +936,20 @@ func (clnt *Client) DropIndex(
 func (clnt *Client) Truncate(policy *WritePolicy, namespace, set string, beforeLastUpdate *time.Time) error {
 	policy = clnt.getUsableWritePolicy(policy)
 
+	node, err := clnt.cluster.GetRandomNode()
+	if err != nil {
+		return err
+	}
+
+	node.tendConnLock.Lock()
+	defer node.tendConnLock.Unlock()
+
+	if err := node.initTendConn(policy.Timeout); err != nil {
+		return err
+	}
+
 	var strCmd bytes.Buffer
-	_, err := strCmd.WriteString("truncate:namespace=")
+	_, err = strCmd.WriteString("truncate:namespace=")
 	_, err = strCmd.WriteString(namespace)
 
 	if len(set) > 0 {
@@ -949,11 +959,15 @@ func (clnt *Client) Truncate(policy *WritePolicy, namespace, set string, beforeL
 	if beforeLastUpdate != nil {
 		_, err = strCmd.WriteString(";lut=")
 		_, err = strCmd.WriteString(strconv.FormatInt(beforeLastUpdate.UnixNano(), 10))
+	} else {
+		if node.supportsLUTNow.Get() {
+			_, err = strCmd.WriteString(";lut=now")
+		}
 	}
 
-	// Send index command to one node. That node will distribute the command to other nodes.
-	responseMap, err := clnt.sendInfoCommand(policy.Timeout, strCmd.String())
+	responseMap, err := RequestInfo(node.tendConn, strCmd.String())
 	if err != nil {
+		node.tendConn.Close()
 		return err
 	}