+++ node.go
@@ -15,6 +15,8 @@
 package aerospike
 
 import (
+	"bufio"
+	"io"
 	"strconv"
 	"strings"
 	"sync"
@@ -40,6 +42,8 @@ type Node struct {
 	_sessionToken      atomic.Value //[]byte
 	_sessionExpiration atomic.Value //time.Time
 
+	racks atomic.Value //map[string]int
+
 	// tendConn reserves a connection for tend so that it won't have to
 	// wait in queue for connections, since that will cause starvation
 	// and the node being dropped under load.
@@ -49,7 +53,7 @@ type Node struct {
 	peersGeneration AtomicInt
 	peersCount      AtomicInt
 
-	connections     connectionQueue //AtomicQueue //ArrayBlockingQueue<*Connection>
+	connections     connectionHeap
 	connectionCount AtomicInt
 	health          AtomicInt //AtomicInteger
 
@@ -61,7 +65,7 @@ type Node struct {
 
 	active AtomicBool
 
-	supportsFloat, supportsBatchIndex, supportsReplicasAll, supportsReplicas, supportsGeo, supportsPeers AtomicBool
+	supportsFloat, supportsBatchIndex, supportsReplicasAll, supportsReplicas, supportsGeo, supportsPeers, supportsLUTNow AtomicBool
 }
 
 // NewNode initializes a server node with connection parameters.
@@ -74,7 +78,7 @@ func newNode(cluster *Cluster, nv *nodeValidator) *Node {
 
 		// Assign host to first IP alias because the server identifies nodes
 		// by IP address (not hostname).
-		connections:         *newConnectionQueue(cluster.clientPolicy.ConnectionQueueSize), //*NewAtomicQueue(cluster.clientPolicy.ConnectionQueueSize),
+		connections:         *newConnectionHeap(cluster.clientPolicy.ConnectionQueueSize),
 		connectionCount:     *NewAtomicInt(0),
 		peersGeneration:     *NewAtomicInt(-1),
 		partitionGeneration: *NewAtomicInt(-2),
@@ -89,10 +93,12 @@ func newNode(cluster *Cluster, nv *nodeValidator) *Node {
 		supportsReplicas:    *NewAtomicBool(nv.supportsReplicas),
 		supportsGeo:         *NewAtomicBool(nv.supportsGeo),
 		supportsPeers:       *NewAtomicBool(nv.supportsPeers),
+		supportsLUTNow:      *NewAtomicBool(nv.supportsLUTNow),
 	}
 
 	newNode.aliases.Store(nv.aliases)
 	newNode._sessionToken.Store(nv.sessionToken)
+	newNode.racks.Store(map[string]int{})
 
 	// this will reset to zero on first aggregation on the cluster,
 	// therefore will only be counted once.
@@ -114,8 +120,10 @@ func (nd *Node) Refresh(peers *peers) error {
 
 	nd.referenceCount.Set(0)
 
+	var infoMap map[string]string
+	var err error
 	if peers.usePeers.Get() {
-		infoMap, err := nd.RequestInfo("node", "peers-generation", "partition-generation")
+		infoMap, err = nd.RequestInfo("node", "peers-generation", "partition-generation", "racks:")
 		if err != nil {
 			nd.refreshFailed(err)
 			return err
@@ -136,9 +144,9 @@ func (nd *Node) Refresh(peers *peers) error {
 			return err
 		}
 	} else {
-		commands := []string{"node", "partition-generation", nd.cluster.clientPolicy.servicesString()}
+		commands := []string{"node", "partition-generation", "racks:", nd.cluster.clientPolicy.servicesString()}
 
-		infoMap, err := nd.RequestInfo(commands...)
+		infoMap, err = nd.RequestInfo(commands...)
 		if err != nil {
 			nd.refreshFailed(err)
 			return err
@@ -159,6 +167,17 @@ func (nd *Node) Refresh(peers *peers) error {
 			return err
 		}
 	}
+
+	if err := nd.updateRackInfo(infoMap); err != nil {
+		// Update rack info should fail if the feature is not supported on the server
+		if aerr, ok := err.(AerospikeError); ok && aerr.ResultCode() == UNSUPPORTED_FEATURE {
+			nd.refreshFailed(err)
+			return err
+		}
+		// Should not fail in other cases
+		Logger.Warn("Updating node rack info failed with error: %s (racks: `%s`)", err, infoMap["racks:"])
+	}
+
 	nd.failures.Set(0)
 	peers.refreshCount.IncrementAndGet()
 	nd.referenceCount.IncrementAndGet()
@@ -174,7 +193,7 @@ func (nd *Node) Refresh(peers *peers) error {
 // refreshSessionToken refreshes the session token if it has been expired
 func (nd *Node) refreshSessionToken() error {
 	// no session token to refresh
-	if len(nd.cluster.clientPolicy.User) == 0 || nd.cluster.clientPolicy.AuthMode != AuthModeExternal {
+	if !nd.cluster.clientPolicy.RequiresAuthentication() || nd.cluster.clientPolicy.AuthMode != AuthModeExternal {
 		return nil
 	}
 
@@ -196,7 +215,7 @@ func (nd *Node) refreshSessionToken() error {
 	}
 
 	command := NewLoginCommand(nd.tendConn.dataBuffer)
-	if err := command.Login(&nd.cluster.clientPolicy, nd.tendConn); err != nil {
+	if err := command.login(&nd.cluster.clientPolicy, nd.tendConn, nd.cluster.Password()); err != nil {
 		// Socket not authenticated. Do not put back into pool.
 		nd.tendConn.Close()
 		return err
@@ -204,6 +223,69 @@ func (nd *Node) refreshSessionToken() error {
 
 	nd._sessionToken.Store(command.SessionToken)
 	nd._sessionExpiration.Store(command.SessionExpiration)
+
+	return nil
+}
+
+func (nd *Node) updateRackInfo(infoMap map[string]string) error {
+	if !nd.cluster.clientPolicy.RackAware {
+		return nil
+	}
+
+	// Do not raise an error if the server does not support rackaware
+	if strings.HasPrefix(strings.ToUpper(infoMap["racks:"]), "ERROR") {
+		return NewAerospikeError(UNSUPPORTED_FEATURE, "You have set the ClientPolicy.RackAware = true, but the server does not support this feature.")
+	}
+
+	ss := strings.Split(infoMap["racks:"], ";")
+	racks := map[string]int{}
+	for _, s := range ss {
+		in := bufio.NewReader(strings.NewReader(s))
+		_, err := in.ReadString('=')
+		if err != nil {
+			return err
+		}
+
+		ns, err := in.ReadString(':')
+		if err != nil {
+			return err
+		}
+
+		for {
+			_, err = in.ReadString('_')
+			if err != nil {
+				return err
+			}
+
+			rackStr, err := in.ReadString('=')
+			if err != nil {
+				return err
+			}
+
+			rack, err := strconv.Atoi(rackStr[:len(rackStr)-1])
+			if err != nil {
+				return err
+			}
+
+			nodesList, err := in.ReadString(':')
+			if err != nil && err != io.EOF {
+				return err
+			}
+
+			nodes := strings.Split(strings.Trim(nodesList, ":"), ",")
+			for i := range nodes {
+				if nodes[i] == nd.name {
+					racks[ns[:len(ns)-1]] = rack
+				}
+			}
+
+			if err == io.EOF {
+				break
+			}
+		}
+	}
+
+	nd.racks.Store(racks)
 
 	return nil
 }
@@ -711,4 +793,16 @@ func (nd *Node) sessionToken() []byte {
 		return st.([]byte)
 	}
 	return nil
+}
+
+// Rack returns the rack number for the namespace.
+func (nd *Node) Rack(namespace string) (int, error) {
+	racks := nd.racks.Load().(map[string]int)
+	v, exists := racks[namespace]
+
+	if exists {
+		return v, nil
+	}
+
+	return -1, newAerospikeNodeError(nd, RACK_NOT_DEFINED)
 }